added testing

+++ queries/deleteUser.js
@@ -6,6 +6,7 @@ deleteUser()
 - @params: 
     dbconn, the connection to the database,
     data, the user provided contact information to be deleted
+    userID, id of user who's contacts we wish to manipulate
 
 - will remove contact(s) from contacts table if present
 
@@ -13,12 +14,10 @@ deleteUser()
 
 */
 
-async function deleteUser( dbconn, data ){
+async function deleteUser( dbconn, data, userId ){
 
-    let del_contacts = 'DELETE FROM contacts WHERE phoneNum=?';
-
-    //let add_contacts = "INSERT INTO contacts (firstName, lastName, phoneNum, fk_contact_num) VALUES (?,?,?, (SELECT phoneNum FROM users WHERE phoneNum=11111111110))";
-    await dbconn.query(del_contacts, [data.contacts.phoneNum], ( res, err ) => {
+    let del_contacts = 'DELETE FROM contacts WHERE phoneNum=? AND userID=?';
+    await dbconn.query(del_contacts, [data, userId], ( res, err ) => {
         if(err){
             throw err;
         }
